requirements.txt
aiobotocore==2.9.0
aiohttp==3.9.1
aiohttp-retry==2.8.3
aioitertools==0.11.0
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
appdirs==1.4.4
async-timeout==4.0.3
asyncssh==2.14.2
atpublic==4.0
attrs==23.2.0
backports.zoneinfo==0.2.1
billiard==4.2.0
blis==0.7.11
boto3==1.33.13
botocore==1.33.13
cachetools==5.3.2
catalogue==2.0.10
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
clean-text==0.6.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
configobj==5.0.8
cryptography==41.0.7
curated-tokenizers==0.0.8
curated-transformers==0.1.1
cymem==2.0.8
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.1.6
dulwich==0.21.7
dvc==3.38.1
dvc-data==3.7.0
dvc-http==2.32.0
dvc-objects==3.0.5
dvc-render==1.0.0
dvc-s3==3.0.1
dvc-studio-client==0.18.0
dvc-task==0.3.0
emoji==1.7.0
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.7.3/en_core_web_trf-3.7.3-py3-none-any.whl#sha256=f72abb34bdf174876bd4267b29b2501677e605e0a251fdc56c163003182ed68b
entrypoints==0.4
exceptiongroup==1.2.0
filelock==3.13.1
flatten-dict==0.4.2
flufl.lock==7.1.1
frozenlist==1.4.1
fsspec==2023.12.2
ftfy==6.1.3
funcy==2.0
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.15.0
google-auth==2.26.2
google-cloud-aiplatform==1.39.0
google-cloud-bigquery==3.15.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.11.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grandalf==0.8
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-status==1.60.0
gto==1.6.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
hydra-core==1.3.2
idna==3.6
importlib-resources==6.1.1
iterative-telemetry==0.0.8
Jinja2==3.1.3
jmespath==1.0.1
kombu==5.3.4
langcodes==3.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
nameparser==1.1.3
networkx==3.1
numpy==1.24.4
omegaconf==2.3.0
openai==1.7.1
orjson==3.9.10
packaging==23.2
pandas==2.0.3
pathspec==0.12.1
platformdirs==3.11.0
preshed==3.0.9
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.7
py4j==0.10.9.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydot==2.0.0
pygit2==1.13.3
Pygments==2.17.2
pygtrie==2.5.0
pyparsing==3.1.1
pyspark==3.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
s3fs==2023.12.2
s3transfer==0.8.2
scmrepo==2.0.2
semver==3.0.2
shapely==2.0.2
shortuuid==1.0.11
shtab==1.6.5
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
spacy==3.7.2
spacy-curated-transformers==0.2.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqltrie==0.11.0
srsly==2.4.8
sympy==1.12
tabulate==0.9.0
thinc==8.2.2
tomlkit==0.12.3
torch==2.1.2
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
Unidecode==1.3.8
urllib3==1.26.18
vine==5.1.0
voluptuous==0.14.1
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
wrapt==1.16.0
yarl==1.9.4
zc.lockfile==3.0.post1
zipp==3.17.0

transformers~=4.38.1